+++ tests/Web/Slack/Experimental/BlocksSpec.hs
@@ -1,59 +1,11 @@
 module Web.Slack.Experimental.BlocksSpec where
 
-import Control.Monad.Writer.Strict
 import Data.StringVariants (mkNonEmptyText)
 import JSONGolden (oneGoldenTestEncode)
 import TestImport
 import Web.Slack.Experimental.Blocks
+import Web.Slack.Experimental.Blocks.Builder (BlockKitBuilderMessage (..), actionsBlock, contextBlock, dividerBlock, headerBlock, runBlockBuilder, sectionBlock, sectionBlockWithAccessory, sectionBlockWithFields)
 import Web.Slack.Experimental.Blocks.Types
-
--- FIXME(jadel, violet): These builder things should probably be put into a
--- module to make them available, but that's later work
-
-headerBlock :: SlackText -> WriterT [SlackBlock] Identity ()
-headerBlock = tell . pure . SlackBlockHeader . SlackPlainTextOnly
-
-sectionBlock :: SlackText -> WriterT [SlackBlock] Identity ()
-sectionBlock = tell . pure . SlackBlockSection . slackSectionWithText
-
-sectionBlockWithFields :: SlackText -> [SlackText] -> WriterT [SlackBlock] Identity ()
-sectionBlockWithFields text fields =
-  tell
-    $ pure
-    $ SlackBlockSection
-    $ SlackSection
-      { slackSectionText = Just text
-      , slackSectionBlockId = Nothing
-      , slackSectionFields = Just fields
-      , slackSectionAccessory = Nothing
-      }
-
-sectionBlockWithAccessory ::
-  SlackText ->
-  SlackAction ->
-  WriterT [SlackBlock] Identity ()
-sectionBlockWithAccessory t b =
-  tell
-    $ pure
-    $ SlackBlockSection
-      SlackSection
-        { slackSectionText = Just t
-        , slackSectionBlockId = Nothing
-        , slackSectionFields = Nothing
-        , slackSectionAccessory = Just $ SlackButtonAccessory b
-        }
-
-dividerBlock :: WriterT [SlackBlock] Identity ()
-dividerBlock = tell . pure $ SlackBlockDivider
-
-contextBlock :: [SlackContent] -> WriterT [SlackBlock] Identity ()
-contextBlock = tell . pure . SlackBlockContext . SlackContext
-
-actionsBlock ::
-  Maybe SlackBlockId ->
-  SlackActionList ->
-  WriterT [SlackBlock] Identity ()
-actionsBlock a b = tell . pure $ SlackBlockActions a b
 
 slackActionDoNothing :: SlackActionId
 slackActionDoNothing = SlackActionId . fromJust . mkNonEmptyText $ "doNothing"
@@ -67,7 +19,8 @@ slackActionDoNothing3 = SlackActionId . fromJust . mkNonEmptyText $ "doNothing3"
 slackShowBlockFormat :: BlockKitBuilderMessage
 slackShowBlockFormat =
   BlockKitBuilderMessage
-    $ execWriter
+    $ runIdentity
+    . runBlockBuilder
     $ do
       headerBlock ("Blah: " <> list ["a", "b", "c"])
       dividerBlock
@@ -104,14 +57,6 @@ slackShowBlockFormat =
     textMessage = message @Text
     list x = textMessage (intercalate ", " x)
     google = OptionalSetting . mkNonEmptyText @3000 $ "https://google.com"
-
--- | This is used so that we can paste the output of these tests directly into
--- the block kit builder:
--- <https://app.slack.com/block-kit-builder>
-newtype BlockKitBuilderMessage = BlockKitBuilderMessage {blocks :: [SlackBlock]}
-
-instance ToJSON BlockKitBuilderMessage where
-  toJSON BlockKitBuilderMessage {blocks} = object [("blocks", toJSON blocks)]
 
 spec :: Spec
 spec = describe "Slack block builder" do